android/src/androidTest/java/com/raulh82vlc/AvengersComics/ui/activities/ComicsListActivityTest.java
/*
 * Copyright (C) 2017 Raul Hernandez Lopez @raulh82vlc
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.raulh82vlc.AvengersComics.ui.activities;

import android.support.test.rule.ActivityTestRule;
import android.support.test.runner.AndroidJUnit4;

import org.junit.After;
import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;

import static org.junit.Assert.assertNotEquals;

/**
 * @author Raul Hernandez Lopez.
 */
@RunWith(AndroidJUnit4.class)
public class ComicsListActivityTest {
    @Rule
    public ActivityTestRule<ComicsListActivity> mActivityTestRule =
            new ActivityTestRule<>(ComicsListActivity.class);

    @Before
    public void setUp() throws Exception {

    }

    @After
    public void tearDown() throws Exception {

    }

    @Test
    public void useLayoutId() throws Exception {
        ComicsListActivity activity1 = mActivityTestRule.getActivity();
        int id = activity1.getLayoutId();

        assertNotEquals(0, id);
    }
}